requirements.txt
flask
boto3
PyYAML